requirements.txt
Flask==3.0.3
psycopg[binary]>=3.2.0
python-dotenv==1.0.1
